+++ config/routes.rb
@@ -7,12 +7,6 @@ Rails.application.routes.draw do
 
   root to: 'home#index'
 
-  get 'account_confirmations/index', to: 'account_confirmation#index'
-  put 'account_confirmations', to: 'account_confirmation#approve_all', as: 'approve_all'
-  put 'account_confirmations/:id', to: 'account_confirmation#approve', as: 'approve_user'
-  delete 'account_confirmations', to: 'account_confirmation#block_all'
-  delete 'account_confirmations/:id', to: 'account_confirmation#block', as: 'block_user'
-
   devise_for :users,
              controllers: {
                omniauth_callbacks: 'users/omniauth_callbacks',
@@ -67,6 +61,10 @@ Rails.application.routes.draw do
 
   # File Store
   get 'file_store' => 'file_store#index'
+
+  namespace :admin do
+    resources :users
+  end
 
   # Sidekiq monitoring
   authenticate :user, ->(u) { u.admin? } do